Change the method of checking for palindrome strings
After extracting all valid letters from the string, make a duplicate
copy of the string array, and reverse that copy. Then join the strings
in the arrays to form two strings; one reversed and one original without
punctuation or spaces. Then compare the two to see if they are matching